+++ App/backend/src/main/java/com/example/backend/common/minio/MinioConfig.java
@@ -1,0 +1,49 @@
+package com.example.backend.common.minio;
+
+import io.minio.BucketExistsArgs;
+import io.minio.MakeBucketArgs;
+import io.minio.MinioClient;
+import lombok.Data;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.stereotype.Component;
+
+@Configuration
+public class MinioConfig {
+
+    public static final String BUCKET_NAME = "sample";
+
+    @Value("${minio.bucket}")
+    private String bucket;
+
+    @Value("${minio.url}")
+    private String url;
+
+    @Value("${minio.accessKey}")
+    private String accessKey;
+
+    @Value("${minio.secretKey}")
+    private String secretKey;
+
+    @Bean
+    public MinioClient minioClient() throws Exception {
+
+        MinioClient minioClient = MinioClient.builder()
+                .endpoint(url)
+                .credentials(accessKey, secretKey)
+                .build();
+
+        if(!minioClient.bucketExists(BucketExistsArgs.builder().bucket(BUCKET_NAME).build())) {
+
+            minioClient.makeBucket(
+                    MakeBucketArgs.builder()
+                            .bucket(BUCKET_NAME)
+                            .build()
+            );
+        }
+
+        return minioClient;
+    }
+
+}
