+++ App/backend/src/main/java/com/example/backend/domain/user/controller/UserController.java
@@ -4,12 +4,10 @@ import com.example.backend.domain.user.dto.UserSignUpDTO;
 import com.example.backend.domain.user.entity.User;
 import com.example.backend.domain.user.repository.UserRepository;
 import com.example.backend.domain.user.service.UserService;
-import com.example.backend.global.exception.dto.ResponseDTO;
-import com.example.backend.global.exception.response.ResponseCodeMessage;
-import com.example.backend.global.jwt.service.JwtService;
+import com.example.backend.global.result.ResultCodeMessage;
+import com.example.backend.global.result.ResultResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -21,11 +19,10 @@ public class UserController {
 
     @PostMapping("/sign-up")
     @ResponseBody
-    public ResponseEntity<ResponseDTO> signUp(@RequestBody UserSignUpDTO userSignUpDTO) throws Exception {
+    public ResponseEntity<ResultResponseDTO> signUp(@RequestBody UserSignUpDTO userSignUpDTO) throws Exception {
         userService.signUp(userSignUpDTO);
         User user = userRepository.findByUsername(userSignUpDTO.getUsername()).get();
-//        return "가입 성공!";
-        return ResponseEntity.ok(ResponseDTO.of(ResponseCodeMessage.REGISTER_SUCCESS, user));
+        return ResponseEntity.ok(ResultResponseDTO.of(ResultCodeMessage.REGISTER_SUCCESS, user));
     }
