+++ App/backend/src/main/java/com/example/backend/domain/user/exception/UserNotExistedException.java
@@ -1,0 +1,4 @@
+package com.example.backend.domain.user.exception;
+
+public class UserNotExistedException{
+}
