+++ App/backend/src/main/java/com/example/backend/global/authorization/oauth2/handler/OAuth2LoginSuccessHandler.java
@@ -1,0 +1,68 @@
+package com.example.backend.global.authorization.oauth2.handler;
+
+import com.example.backend.domain.user.Enum.Role;
+import com.example.backend.domain.user.entity.User;
+import com.example.backend.domain.user.repository.UserRepository;
+import com.example.backend.global.authorization.jwt.service.JwtService;
+import com.example.backend.global.authorization.oauth2.auth.CustomOAuth2User;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
+
+    private final JwtService jwtService;
+    private final UserRepository userRepository;
+
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
+        log.info("OAuth2 Login 성공!");
+        try {
+            CustomOAuth2User oAuth2User = (CustomOAuth2User) authentication.getPrincipal();
+
+            // User의 Role이 GUEST일 경우 처음 요청한 회원이므로 회원가입 페이지로 리다이렉트
+            if(oAuth2User.getRole() == Role.GUEST) {
+
+                User findUser = userRepository.findByUsername(oAuth2User.getUsername())
+                        .orElseThrow(() -> new IllegalArgumentException("아이디에 해당하는 유저가 없습니다."));
+                String accessToken = jwtService.createAccessToken(oAuth2User.getUsername(), findUser.getNickname());
+                response.addHeader(jwtService.getAccessHeader(), "Bearer " + accessToken);
+                response.sendRedirect("oauth2/sign-up"); // 프론트의 회원가입 추가 정보 입력 폼으로 리다이렉트
+
+                jwtService.sendAccessAndRefreshToken(response, accessToken, null);
+
+//                findUser.authorizeUser();
+            } else {
+                loginSuccess(response, oAuth2User); // 로그인에 성공한 경우 access, refresh 토큰 생성
+            }
+        } catch (Exception e) {
+            throw e;
+        }
+
+    }
+
+    // TODO : 소셜 로그인 시에도 무조건 토큰 생성하지 말고 JWT 인증 필터처럼 RefreshToken 유/무에 따라 다르게 처리해보기
+    private void loginSuccess(HttpServletResponse response, CustomOAuth2User oAuth2User) throws IOException {
+
+        User findUser = userRepository.findByUsername(oAuth2User.getUsername())
+                .orElseThrow(() -> new IllegalArgumentException("아이디에 해당하는 유저가 없습니다."));
+
+        String accessToken = jwtService.createAccessToken(oAuth2User.getUsername(), findUser.getNickname());
+        String refreshToken = jwtService.createRefreshToken();
+        response.addHeader(jwtService.getAccessHeader(), "Bearer " + accessToken);
+        response.addHeader(jwtService.getRefreshHeader(), "Bearer " + refreshToken);
+
+        jwtService.sendAccessAndRefreshToken(response, accessToken, refreshToken);
+        jwtService.updateRefreshToken(oAuth2User.getUsername(), refreshToken);
+    }
+}
