+++ App/backend/src/main/java/com/example/backend/global/authorization/oauth2/userinfo/GoogleOAuth2UserInfo.java
@@ -1,0 +1,20 @@
+package com.example.backend.global.authorization.oauth2.userinfo;
+
+import java.util.Map;
+
+public class GoogleOAuth2UserInfo extends OAuth2UserInfo{
+
+    public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
+        super(attributes);
+    }
+
+    @Override
+    public String getId() {
+        return (String) attributes.get("sub");
+    }
+
+    @Override
+    public String getNickname() {
+        return (String) attributes.get("name");
+    }
+}
