+++ App/backend/src/main/java/com/example/backend/global/authorization/oauth2/userinfo/OAuth2UserInfo.java
@@ -1,0 +1,18 @@
+package com.example.backend.global.authorization.oauth2.userinfo;
+
+import java.util.Map;
+
+public abstract class OAuth2UserInfo {
+
+    protected Map<String, Object> attributes;
+
+    public OAuth2UserInfo(Map<String, Object> attributes) {
+        this.attributes = attributes;
+    }
+
+    public abstract String getId();
+    //소셜 식별 값 : 구글 - "sub", 카카오 - "id", 네이버 - "id"
+
+    public abstract String getNickname();
+
+}
