--- App/backend/src/main/java/com/example/backend/global/login/service/CustomLoginService.java
@@ -1,28 +1,0 @@
-package com.example.backend.global.login.service;
-
-import com.example.backend.domain.user.entity.User;
-import com.example.backend.domain.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class CustomLoginService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername(username) // 해당 아이디를 가진 유저를 찾아낸다.
-                .orElseThrow(() -> new UsernameNotFoundException("해당 아이디가 존재하지 않습니다."));
-
-        return org.springframework.security.core.userdetails.User.builder()
-                .username(user.getUsername())
-                .password(user.getPassword())
-                .roles(user.getRole().name())
-                .build();
-    }
-}
