--- App/backend/src/main/java/com/example/backend/global/oauth2/userinfo/NaverOAuth2UserInfo.java
@@ -1,31 +1,0 @@
-package com.example.backend.global.oauth2.userinfo;
-
-import java.util.Map;
-
-public class NaverOAuth2UserInfo extends OAuth2UserInfo{
-
-    public NaverOAuth2UserInfo(Map<String, Object> attributes) {
-        super(attributes);
-    }
-
-    @Override
-    public String getId() {
-        Map<String, Object> response = (Map<String, Object>) attributes.get("response");
-
-        if (response == null) {
-            return null;
-        }
-        return (String) response.get("id");
-    }
-
-    @Override
-    public String getNickname() {
-        Map<String, Object> response = (Map<String, Object>) attributes.get("response");
-
-        if (response == null) {
-            return null;
-        }
-
-        return (String) response.get("nickname");
-    }
-}
